Move `Validation` to `validate`
This follows the convention set by other layers (both those finished and
those currently being worked on), which all only have the layer-specific
stuff within `layers`, while leaving the rest outside.
