+++ crates/fj-core/src/validate/mod.rs
@@ -76,13 +76,17 @@ mod solid;
 mod surface;
 mod vertex;
 
+use crate::storage::ObjectId;
+
 pub use self::{
     cycle::CycleValidationError, edge::EdgeValidationError,
     face::FaceValidationError, shell::ShellValidationError,
     sketch::SketchValidationError, solid::SolidValidationError,
 };
 
-use std::{convert::Infallible, fmt};
+use std::{
+    collections::HashMap, convert::Infallible, error::Error, fmt, thread,
+};
 
 use fj_math::Scalar;
 
@@ -97,6 +101,59 @@ macro_rules! assert_contains_err {
             errors.iter().any(|e| matches!(e, $p))
         })
     };
+}
+
+/// Errors that occurred while validating the objects inserted into the stores
+#[derive(Default)]
+pub struct Validation {
+    /// All unhandled validation errors
+    pub errors: HashMap<ObjectId, ValidationError>,
+
+    /// Validation configuration for the validation service
+    pub config: ValidationConfig,
+}
+
+impl Validation {
+    /// Construct an instance of `Validation`, using the provided configuration
+    pub fn with_validation_config(config: ValidationConfig) -> Self {
+        let errors = HashMap::new();
+        Self { errors, config }
+    }
+
+    /// Drop this instance, returning the errors it contained
+    pub fn into_errors(mut self) -> ValidationErrors {
+        ValidationErrors(self.errors.drain().map(|(_, error)| error).collect())
+    }
+}
+
+impl Drop for Validation {
+    fn drop(&mut self) {
+        let num_errors = self.errors.len();
+        if num_errors > 0 {
+            println!(
+                "Dropping `Validation` with {num_errors} unhandled validation \
+                errors:"
+            );
+
+            for err in self.errors.values() {
+                println!("{}", err);
+
+                // Once `Report` is stable, we can replace this:
+                // https://doc.rust-lang.org/std/error/struct.Report.html
+                let mut source = err.source();
+                while let Some(err) = source {
+                    println!("\nCaused by:\n\t{err}");
+                    source = err.source();
+                }
+
+                print!("\n\n");
+            }
+
+            if !thread::panicking() {
+                panic!();
+            }
+        }
+    }
 }
 
 /// Validate an object
